refactor: Simplify PasswordStrengthIndicator imports in Landing and Reset Password pages
- Updated imports of `PasswordStrengthIndicator` in `LandingPage` and `ResetPassword` components to streamline the codebase.
- Clarified documentation in `passwordValidation.ts` to better explain password strength calculation rules.
